requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
statsmodels>=0.14
matplotlib>=3.9
pyyaml>=6.0
joblib>=1.4
jupyter>=1.0
ruptures>=1.1
